it...works?  i think?  wah

+++ app/scripts/controllers/room.js
@@ -1,7 +1,8 @@
 'use strict';
 
 angular.module('youKaraokeApp')
-    .controller('RoomCtrl', function($scope, $http, auth, localStorageService, $routeParams, $location, fb) {
+    .controller('RoomCtrl', function($scope, $http, auth, localStorageService, $routeParams, $location, fb, $firebase) {
+    	console.log($firebase);
         localStorageService.set("lastsite", $routeParams.id);
         if (!auth.getCurrentUser()) {
             $location.path('/main');
@@ -63,23 +64,57 @@ angular.module('youKaraokeApp')
 
         videosRef.on('child_added', function(dataSnapshot) {
         	console.log("Videos been added ", dataSnapshot.val());
-            $http({
-                    url: 'https://www.googleapis.com/youtube/v3/videos',
-                    method: 'GET',
-                    params: {
-                        part: 'snippet',
-                        id: dataSnapshot.val(),
-                        maxResults: 4
-                    },
-                    headers: {
-                        Authorization: 'Bearer ' + $scope.currentUser.google.accessToken
+        	if ($scope.isCreator()) {
+            	$scope.player.addEventListener('onStateChange', reloadWhenDone);
+            }
+            // $http({
+            //         url: 'https://www.googleapis.com/youtube/v3/videos',
+            //         method: 'GET',
+            //         params: {
+            //             part: 'snippet',
+            //             id: dataSnapshot.val(),
+            //             maxResults: 4
+            //         },
+            //         headers: {
+            //             Authorization: 'Bearer ' + $scope.currentUser.google.accessToken
+            //         }
+            //     })
+            //     .success(function(res) {
+            //         $scope.queue2 = res;
+            //         console.log("fuck this: ", $scope.queue2.items[0].snippet.title);
+            //     });	
+
+	        $http({
+                url: 'https://www.googleapis.com/youtube/v3/playlistItems',
+                method: 'GET',
+                params: {
+                    part: 'snippet',
+                    playlistId: $scope.playlist[0].id, // whatever the playlist id actually is
+                    maxResults: 50
+                },
+                headers: {
+                    Authorization: 'Bearer ' + $scope.creator.google.accessToken
+                }
+            })
+            .success(function(res) {
+            	console.log(res);
+                $scope.queue = res.items.map(function(item) {
+                    return {
+                        title: item.snippet.title,
+                        id: item.id,
+                        videoId: item.snippet.resourceId.videoId,
+                        status: 'non'
                     }
-                })
-                .success(function(res) {
-                    $scope.queue2 = res;
-                    console.log("fuck this: ", $scope.queue2.items[0].snippet.title);
-                });	
+                });
+                console.log($scope.queue);
+
+                // $scope.queue[$scope.player.playlistIndex()].status = 'current';
+            });
         });
+
+		videosRef.on('child_removed', function(dataSnapshot) {
+			console.log("Video removed ", dataSnapshot.val());
+		})
 
 
         
@@ -88,7 +123,6 @@ angular.module('youKaraokeApp')
 
         //USERS
         var usersRef = fb.room.child($routeParams.id).child("users");
-
 
         usersRef.on('child_added', function(dataSnapshot) {
             console.log("CHILD ADDED TO USERS", dataSnapshot.val().google.displayName);
@@ -109,23 +143,66 @@ angular.module('youKaraokeApp')
         
 
 
+        if ($scope.isCreator()) {
+	        var tag = document.createElement('script');
+	        tag.src = "https://www.youtube.com/iframe_api";
 
+	        var firstScriptTag = document.getElementsByTagName('script')[0];
+	        firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
 
-        //melissa
-
-        var tag = document.createElement('script');
-        tag.src = "https://www.youtube.com/iframe_api";
-        var firstScriptTag = document.getElementsByTagName('script')[0];
-        firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
+	        $scope.player;
+	        $scope.queue;
+	        $scope.youTubeIframeAPIReady = false;
+	        window.onYouTubeIframeAPIReady = runWhenAPIReady;
+        	
+	        if (window.YT) runWhenAPIReady();
+        }
 
         var done = false;
-        $scope.player;
-        $scope.queue;
-        $scope.youTubeIframeAPIReady = false;
-        window.onYouTubeIframeAPIReady = runWhenAPIReady;
 
-        if (window.YT) runWhenAPIReady();
+/***** RELOADING EVENT LISTENER.  GETS ADDED UPON CHILD ADD *****/
 
+	  	var reloadWhenDone = function(evt) {
+            if (evt.data == 0) {
+                if (typeof $scope.player.getPlaylistIndex() !== 'undefined') {
+                    index = $scope.player.getPlaylistIndex();
+                    $scope.player.loadPlaylist({
+                        listType: 'playlist',
+                        list: $scope.playlist[0].id,
+                        index: index
+                    });
+                }
+                $scope.player.removeEventListener('onStateChange', reloadWhenDone);
+            }
+        }
+
+/***** POPULATE QUEUE IMMEDIATELY *****/
+
+       $http({
+            url: 'https://www.googleapis.com/youtube/v3/playlistItems',
+            method: 'GET',
+            params: {
+                part: 'snippet',
+                playlistId: $scope.playlist[0].id,
+                maxResults: 50
+            },
+            headers: {
+                Authorization: 'Bearer ' + $scope.currentUser.google.accessToken
+            }
+        })
+        .success(function(res) {
+            console.log(res);
+            $scope.queue = res.items.map(function(item) {
+                return {
+                    title: item.snippet.title,
+                    id: item.id,
+                    status: 'non'
+                }
+            });
+            $scope.queue[0].status = 'current';
+        });
+
+/***** PLAYER FUNCTIONS *****/
         function runWhenAPIReady() {
             $scope.youTubeIframeAPIReady = window.youTubeIframeAPIReady = true;
             $scope.player = new YT.Player('player', {
@@ -138,39 +215,14 @@ angular.module('youKaraokeApp')
                     'onReady': onPlayerReady,
                     'onStateChange': onPlayerStateChange
                 }
-                //optional: playerVars
             });
         };
 
         function onPlayerReady(evt) {
             $scope.player.loadPlaylist({
                 listType: 'playlist',
-                list: $scope.playlist[0].id // whatever the playlist id actually is
+                list: $scope.playlist[0].id
             });
-            // evt.target.playVideo();
-            $http({
-                    url: 'https://www.googleapis.com/youtube/v3/playlistItems',
-                    method: 'GET',
-                    params: {
-                        part: 'snippet',
-                        playlistId: $scope.playlist[0].id, // whatever the playlist id actually is
-                        maxResults: 50
-                    },
-                    headers: {
-                        Authorization: 'Bearer ' + $scope.currentUser.google.accessToken
-                    }
-                })
-                .success(function(res) {
-                    console.log(res);
-                    $scope.queue = res.items.map(function(item) {
-                        return {
-                            title: item.snippet.title,
-                            id: item.id,
-                            status: 'non'
-                        }
-                    });
-                    $scope.queue[0].status = 'current';
-                });
         };
 
         function onPlayerStateChange(evt) {
@@ -180,7 +232,6 @@ angular.module('youKaraokeApp')
             }
             if (evt.data == 0) {
                 var i = $scope.player.getPlaylistIndex();
-                // shhhhhh it's going to be okay
                 if ($scope.queue[i - 1].status !== 'non' || i === 1) {
                     $scope.queue[i - 1].status = 'non';
                     $scope.queue[i].status = 'current';
@@ -192,6 +243,8 @@ angular.module('youKaraokeApp')
         function stopVideo() {
             $scope.player.stopVideo();
         };
+
+/***** SEARCH *****/
 
         $scope.searchResults = [];
         $scope.search = function(query) {
@@ -214,24 +267,11 @@ angular.module('youKaraokeApp')
         };
 
         var index = 0;
-        var reloadWhenDone = function(evt) {
-            if (evt.data == 0) {
-                // don't even look at me
-                console.log("WE MADE IT HEREREERERER");
-                if (typeof $scope.player.getPlaylistIndex() !== 'undefined') {
-                    index = $scope.player.getPlaylistIndex();
-                    $scope.player.loadPlaylist({
-                        listType: 'playlist',
-                        list: $scope.playlist[0].id, // whatever the playlist id actually is
-                        index: index
-                    });
-                }
-                $scope.player.removeEventListener('onStateChange', reloadWhenDone);
-            }
-        }
+
+/***** ADD TO PLAYLIST--FOR USERS ALSO *****/
 
         $scope.addToPlaylist = function(videoId) {
-        	videosRef.push(videoId);
+        	console.log($scope.creator.google.accessToken);
             $http({
                     url: 'https://www.googleapis.com/youtube/v3/playlistItems',
                     method: 'POST',
@@ -255,54 +295,44 @@ angular.module('youKaraokeApp')
                     angular.element('#' + videoId).css({
                         display: 'block'
                     });
-                    
-                    $scope.player.addEventListener('onStateChange', reloadWhenDone);
-                    
-                    $http({
-                            url: 'https://www.googleapis.com/youtube/v3/playlistItems',
-                            method: 'GET',
-                            params: {
-                                part: 'snippet',
-                                playlistId: $scope.playlist[0].id, // whatever the playlist id actually is
-                                maxResults: 50
-                            },
-                            headers: {
-                                Authorization: 'Bearer ' + $scope.creator.google.accessToken
-                            }
-                        })
-                        .success(function(res) {
-	                            $scope.queue = res.items.map(function(item) {
-	                                return {
-	                                    title: item.snippet.title,
-	                                    id: item.id,
-	                                    status: 'non'
-	                                }
-	                            });
-                            $scope.queue[$scope.player.getPlaylistIndex()].status = 'current';
-                        });
-                        
+		        	videosRef.push(videoId);
                 });
         };
 
-        $scope.removeFromPlaylist = function(id) {
+/***** REMOVE FROM PLAYLIST - FOR CREATOR ONLY *****/
+
+		$scope.removeFromPlaylist = function(video) {
+			if (!$scope.isCreator()) return false;
             $http({
                     url: 'https://www.googleapis.com/youtube/v3/playlistItems',
                     method: 'DELETE',
                     params: {
-                        id: id
+                        id: video.id
                     },
                     headers: {
                         Authorization: 'Bearer ' + $scope.creator.google.accessToken
                     }
                 })
                 .success(function(res) {
+                	var videosArr = $firebase(videosRef).$asArray();
+                	videosArr.$loaded().then(function(){
+                		videosArr.forEach(function(item) {
+                			if (item.$value == video.videoId) {
+			                	var videoChildRef = fb.room.child($routeParams.id).child("videos").child(item.$id);
+                				videoChildRef.remove();
+                			}
+                		})
+                	})
+                	// console.log(videosArr[0]);
+                	// videoChildRef.remove();
+
                     $scope.player.addEventListener('onStateChange', reloadWhenDone);
                     $http({
                             url: 'https://www.googleapis.com/youtube/v3/playlistItems',
                             method: 'GET',
                             params: {
                                 part: 'snippet',
-                                playlistId: $scope.playlist[0].id, // whatever the playlist id actually is
+                                playlistId: $scope.playlist[0].id,
                                 maxResults: 50
                             },
                             headers: {
@@ -321,4 +351,5 @@ angular.module('youKaraokeApp')
                         });
                 });
         }
+        
     });
